requirements.txt
gradio
google-generativeai
PyPDF2